+++ airbyte-config/init/src/main/resources/seed/source_definitions.yaml
@@ -124,10 +124,10 @@
   dockerImageTag: 0.1.0
   documentationUrl: https://https://docs.airbyte.io/integrations/sources/greenhouse
 - sourceDefinitionId: d29764f8-80d7-4dd7-acbe-1a42005ee5aa
-  name: Zendesk
-  dockerRepository: airbyte/source-zendesk-singer
+  name: Zendesk Support
+  dockerRepository: airbyte/source-zendesk-support-singer
   dockerImageTag: 0.1.1
-  documentationUrl: https://hub.docker.com/r/airbyte/source-zendesk-singer
+  documentationUrl: https://hub.docker.com/r/airbyte/source-zendesk-support-singer
 - sourceDefinitionId: d8313939-3782-41b0-be29-b3ca20d8dd3a
   name: Intercom
   dockerRepository: airbyte/source-intercom-singer
