+++ airbyte-integrations/connectors/source-intercom-singer/unit_tests/unit_test.py
@@ -22,6 +22,5 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 """
 
-
 def test_example_method():
     assert True
