+++ airbyte-integrations/connectors/source-zendesk-support-singer/integration_tests/catalog_subset.json
