+++ airbyte-integrations/connectors/source-zendesk-support-singer/sample_files/sample_catalog.json
