+++ airbyte-integrations/connectors/source-zendesk-support-singer/source_zendesk_singer/__init__.py
@@ -22,6 +22,6 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 """
 
-from .source import SourceZendeskSinger
+from .source import SourceZendeskSupportSinger
 
-__all__ = ["SourceZendeskSinger"]
+__all__ = ["SourceZendeskSupportSinger"]
