+++ airbyte-integrations/connectors/source-zendesk-support-singer/source_zendesk_singer/spec.json
